put execution part into a separate file
