+++ hierarchical_LTL.py
@@ -421,7 +421,7 @@ def vis_graph(graph, att, title, latex=False):
     # add the following the generated dot file
     # rankdir=LR;
 	# node [texmode="math"];
-    #  dot2tex --preproc ./data/task_network.dot | dot2tex > ./data/task_network.tex
+    # dot2tex --preproc --texmode math ./data/task_network.dot | dot2tex > ./data/task_network.tex
     if not latex:
         # Run a Linux command
         command = "dot -Tpng {0}.dot >{0}.png".format(title)
@@ -466,86 +466,6 @@ def print_timed_plan(robot_time, robot_waypoint, robot_label, time_axis, robot_t
     
 def get_finished_task():
     return -1
-
-def task_execution(time_task_element_type_robot_axis, reduced_task_network, type_robots):
-    finished_tasks = []
-    current_exec_tasks = []
-    current_exec_robots = []
-    candidate_tasks = set()
-    
-    task_element2type_robot = {time_task_element_type_robot[1] : time_task_element_type_robot[2] \
-        for time_task_element_type_robot in time_task_element_type_robot_axis if len(time_task_element_type_robot) == 3}
-    task_element2time =  {time_task_element_type_robot[1] : time_task_element_type_robot[0] \
-        for time_task_element_type_robot in time_task_element_type_robot_axis if len(time_task_element_type_robot) == 3}
-    print("task_element2type_robot: ", task_element2type_robot)
-    print("task_element2time: ", task_element2time)
-    executing_task_network = reduced_task_network.copy()
-    
-    invalid_str = "-1"
-    finished_task_str = invalid_str
-    while executing_task_network.nodes():
-        # candidate tasks: 1. no parent in updated task network 2. no parent in executing tasks
-        candidate_tasks = set(node for node in executing_task_network.nodes() \
-            if executing_task_network.in_degree(node) == 0)
-        removed_candidate_tasks = set(task for task in candidate_tasks for exec_task in current_exec_tasks \
-            if (exec_task, task) in executing_task_network.edges())
-        removed_candidate_tasks.update(removed_candidate_tasks)
-        candidate_tasks.difference_update(removed_candidate_tasks)
-        # deal with subtasks that are ignored due to or relation
-        remove_task_element_from_network = set(task for task in candidate_tasks if task not in task_element2type_robot.keys())
-        candidate_tasks.difference_update(remove_task_element_from_network)
-        executing_task_network.remove_nodes_from(remove_task_element_from_network)
-        
-        for type_robot in type_robots:
-            if type_robot in current_exec_robots:
-                continue
-            candidate_tasks_for_robot = [task_element for task_element in candidate_tasks \
-                if task_element2type_robot[task_element] == type_robot]
-            if not candidate_tasks_for_robot:
-                continue
-            # if there are several candiate tasks for a robot, sort according to task plan
-            ordered_candidata_tasks = [task_element2time[task_element] for task_element in candidate_tasks_for_robot]
-            task_index = ordered_candidata_tasks.index(min(ordered_candidata_tasks))
-
-            current_exec_robots.append(type_robot)
-            current_exec_tasks.append(candidate_tasks_for_robot[task_index])
-            
-            # generate message 
-            is_human = type_robot == (1, 1)
-            task_id = reduced_task_network.nodes[current_exec_tasks[-1]]['label'][0][0][0]
-            print("is human {0}, task id {1}".format(is_human, task_id))
-        # remove current subtasks
-        executing_task_network.remove_nodes_from(current_exec_tasks)
-        print("current_exec_robots: ", current_exec_robots)
-        for task in current_exec_tasks:
-            print("current_exec_tasks: ", task, reduced_task_network.nodes[task]['label'], 
-                  reduced_task_network.nodes[task]['label'][0][0][0])
-        
-        # if finished_task_str == invalid_str:
-        #     finished_task_str = input("Finished task: ")
-        #     finished_task_split = finished_task_str.split()     
-        #     finished_task_str = invalid_str
-        #     # update when some tasks are finished
-        #     finished_task = (finished_task_split[0], int(finished_task_split[1]))
-        #     finished_tasks.append(finished_task)
-        #     print("finished_task: ", finished_task)
-        #     current_exec_tasks.remove(finished_task)
-        #     current_exec_robots.remove(task_element2type_robot[finished_task])
-            
-        # (1, 1) human (1, 0) robot
-        while finished_task_str == invalid_str:
-            finished_task_str = input("Finished task: ")
-            if finished_task_str != '0' and finished_task_str != '1':
-                finished_task_str = invalid_str
-        is_human = int(finished_task_str) == 1
-        finished_task_str = invalid_str
-        if is_human:
-            task_idx = current_exec_robots.index((1,1))
-        else:
-            task_idx = current_exec_robots.index((1,0))
-        print("finished_task: ", current_exec_tasks[task_idx])
-        current_exec_tasks.pop(task_idx)
-        current_exec_robots.pop(task_idx)
             
 def generate_latex_expr(label, task):
     if task == "nav":
@@ -578,25 +498,10 @@ def generate_latex_expr(label, task):
         final_expression = f"\\top"
     return final_expression
 
-
-def create_parser():
-    """ create parser
-
-    Returns:
-        _type_: _description_
-    """
-    parser = argparse.ArgumentParser(description='FM')
-    parser.add_argument('--task', default="man", type=str)
-    parser.add_argument('--case', default=0, type=int)
-    parser.add_argument('--vis', action='store_true', help='Enable visualization')
-    parser.add_argument('--dot', action='store_true', help='Enable dot graph')
-
-    return parser
   
-def main():
+def hierarchical_ltl_planner(args):
     start = datetime.now()
-    parser = create_parser()
-    args = parser.parse_known_args()[0]
+
 
     # ----------------- task -----------------
     task_specification = Specification().get_task_specification(args.task, args.case)
@@ -671,14 +576,12 @@ def main():
         plt.savefig('./data/workspace.png') 
 
 
-    if reduced_task_network.graph["task"] == "man":
-        task_execution(time_task_element_type_robot_axis, reduced_task_network, workspace.type_robot_location.keys())
-    else:
+    if reduced_task_network.graph["task"] == "nav":
         # --------------------- GMRPP -------------------------
         robot_path = mapp(workspace, acpt_run, robot_waypoint_axis, robot_time_axis, args.vis)
         cost = compute_path_cost(robot_path)
         print('{1}, {0}\n'.format((datetime.now() - start).total_seconds(), cost))
         if args.vis:
-            vis(workspace, robot_path, {robot: [len(path)] * 2 for robot, path in robot_path.items()}, [])
-if __builtins__:
-    main()
+            vis(args.task, args.case, workspace, robot_path, {robot: [len(path)] * 2 for robot, path in robot_path.items()}, [])
+    else:
+        return time_task_element_type_robot_axis, reduced_task_network, workspace.type_robot_location.keys()
